feat: authorization
Enable authorized users to create and modify bids.

+++ mutations/bidMutations.js
@@ -4,8 +4,18 @@ const {
   deleteBidController,
 } = require("../controllers/bidController");
 
-exports.createBid = function (_, args) {
+const { GraphQLError } = require("graphql");
+
+exports.createBid = function (_, args, { credentials }) {
   const { proposal, bidder_id, job_id } = args;
+
+  if (credentials !== "Worker") {
+    throw new GraphQLError("You are not authorized to create a bid.", {
+      extensions: {
+        code: "FORBIDDEN",
+      },
+    });
+  }
 
   return createBidController({
     proposal: proposal,
@@ -17,11 +27,27 @@ exports.createBid = function (_, args) {
 exports.updateBid = function (_, args) {
   const { id, proposal } = args;
 
+  if (credentials !== "Worker") {
+    throw new GraphQLError("You are not authorized to update a bid.", {
+      extensions: {
+        code: "FORBIDDEN",
+      },
+    });
+  }
+
   return updateBidController({ id: id, proposal: proposal });
 };
 
 exports.deleteBid = function (_, args) {
   const { id } = args;
+
+  if (credentials !== "Worker" || credentials !== "Admin") {
+    throw new GraphQLError("You are not authorized to dekete a bid.", {
+      extensions: {
+        code: "FORBIDDEN",
+      },
+    });
+  }
 
   return deleteBidController({ id: id });
 };
